started routing the PCB
started routing the PCB - some components moved a bit. Added GND fill/pour on top and bottom planes, adding extra GND trace width where needed. Also routed short power traces on the top layer. The USB-C 5V trace is a bit long and might need to be moved to layer 2 at some point, but that would require a 70mil trace instead of a 14mil trace for handling 1.5A on an inner layer that is thinner.